# backend/src/main/resources/openapi/p03.json
{
  "openapi":"3.0.3",
  "info":{"title":"LedgerGuard authentication and accounts API","version":"0.3.0","description":"Synthetic money only. P03 contract: registration, revocable sessions and ownership-safe accounts. Transfers, asynchronous payments and the React UI are later milestones, not implemented HTTP workflows. Authentication uses an HS256 signed JWT in an HttpOnly cookie, backed by a PostgreSQL session checked on every protected request. Default lifetime is 15 minutes; no refresh tokens. Obtain CSRF at startup and again after successful login/logout. Send the returned masked token in X-XSRF-TOKEN; do not read or persist authentication cookies. Normal deployment requires HTTPS with __Host- cookies; the explicit loopback sandbox profile uses LG-SESSION and LG-CSRF. All monetary values and database sequence/version values are decimal integer strings. Lists use stable documented ordering and bounded offset pagination; concurrent new records can shift offsets, so this is not a snapshot cursor."},
  "servers":[{"url":"/"}],
  "security":[{"SessionCookie":[]}],
  "paths":{
    "/api/v1/auth/csrf":{"get":{"operationId":"acquireCsrf","security":[],"description":"Return a BREACH-masked CSRF token and set an HttpOnly CSRF cookie. Call again after login/logout; response must not be cached.","responses":{"200":{"description":"CSRF token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Csrf"}}}}}}},
    "/api/v1/auth/register":{"post":{"operationId":"registerCustomer","security":[{"CsrfHeader":[]}],"description":"Trim and lowercase the email using locale-independent normalization. Create only CUSTOMER; registration does not log in and does not create/fund accounts. Unknown JSON properties are rejected. No administrator role is accepted.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Registration"}}}},"responses":{"201":{"description":"Registered customer","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Registered"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"403":{"$ref":"#/components/responses/Forbidden"},"409":{"description":"Registration rejected, without disclosing user details","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},"429":{"$ref":"#/components/responses/Throttled"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/auth/login":{"post":{"operationId":"login","security":[{"CsrfHeader":[]}],"description":"Validate credentials, persist a session, issue an HttpOnly cookie, then clear the previous CSRF cookie. Reacquire CSRF before another command. IP and identity budgets are shared across API instances; default five-minute windows allow 60 IP attempts and 10 attempts per normalized identity and operation.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"200":{"description":"Authenticated identity; token only in Set-Cookie, never JSON","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"429":{"$ref":"#/components/responses/Throttled"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/auth/me":{"get":{"operationId":"currentUser","responses":{"200":{"description":"Current enabled user and persisted session expiry","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Session"}}}},"401":{"$ref":"#/components/responses/Unauthorized"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/auth/logout":{"post":{"operationId":"logout","security":[{"SessionCookie":[],"CsrfHeader":[]}],"description":"Revoke the persisted session and expire authentication/CSRF cookies. A copied JWT is no longer valid, including on another process or after restart. No request body is needed.","responses":{"204":{"description":"Session revoked; empty body"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/accounts":{"get":{"operationId":"listOwnedAccounts","description":"CUSTOMER only. Ownership is applied in SQL. Order: created_at descending, then immutable id descending.","parameters":[{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Offset"}],"responses":{"200":{"description":"Owned accounts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountPage"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"503":{"$ref":"#/components/responses/Unavailable"}}},"post":{"operationId":"createAccount","security":[{"SessionCookie":[],"CsrfHeader":[]}],"description":"CUSTOMER only. Create an owned zero-balance wallet through the protected database function. Caller cannot assign an owner, balance, account kind, id, status or public reference. No synthetic funding API is exposed.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccount"}}}},"responses":{"201":{"description":"Created account","headers":{"Location":{"schema":{"type":"string"},"description":"Relative account resource URL"}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"413":{"description":"Body exceeds 16384 bytes","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/accounts/{id}":{"parameters":[{"$ref":"#/components/parameters/AccountId"}],"get":{"operationId":"getOwnedAccount","responses":{"200":{"description":"Authoritative owner-scoped balance","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"404":{"$ref":"#/components/responses/NotFound"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/accounts/{id}/entries":{"parameters":[{"$ref":"#/components/parameters/AccountId"},{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Offset"}],"get":{"operationId":"listOwnedEntries","description":"Only this owned account's immutable entry lines, ordered by entry id descending. No counterparty balances, private account IDs or unrelated journal lines.","responses":{"200":{"description":"Owned journal entries","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntryPage"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"404":{"$ref":"#/components/responses/NotFound"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/accounts/{id}/transactions":{"parameters":[{"$ref":"#/components/parameters/AccountId"},{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Offset"}],"get":{"operationId":"listOwnedTransactions","description":"This account's posted economic effects, grouped by journal. Order: journal created_at descending, id descending. Reservations are not posted journal movements.","responses":{"200":{"description":"Owned transaction history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionPage"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"404":{"$ref":"#/components/responses/NotFound"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/recipients/{publicRef}":{"get":{"operationId":"resolveRecipient","description":"CUSTOMER only. Resolve an active wallet of an enabled user to its public reference and currency only. Does not grant account/history access; excludes system funding accounts.","parameters":[{"name":"publicRef","in":"path","required":true,"schema":{"type":"string","pattern":"^LG-[a-fA-F0-9]{32}$"}}],"responses":{"200":{"description":"Minimal routing information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Recipient"}}}},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"404":{"$ref":"#/components/responses/NotFound"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/admin/security-events":{"get":{"operationId":"listSecurityEvents","description":"ADMIN only. Separate append-only authentication/authorization/account-creation event feed, not a financial audit search API. Order: id descending. Contains no credentials or request payloads. Administrator status does not allow customer account creation or unrestricted spending.","parameters":[{"$ref":"#/components/parameters/Limit"},{"$ref":"#/components/parameters/Offset"}],"responses":{"200":{"description":"Security events","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SecurityEventPage"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"401":{"$ref":"#/components/responses/Unauthorized"},"403":{"$ref":"#/components/responses/Forbidden"},"503":{"$ref":"#/components/responses/Unavailable"}}}},
    "/api/v1/openapi.json":{"get":{"operationId":"openApi","security":[],"responses":{"200":{"description":"This versioned OpenAPI document","content":{"application/json":{"schema":{"type":"object"}}}}}}}
  },
  "components":{
    "securitySchemes":{"SessionCookie":{"type":"apiKey","in":"cookie","name":"__Host-LG-SESSION","description":"HttpOnly, Secure, SameSite=Strict, Path=/, no Domain. Explicit localhost sandbox uses LG-SESSION."},"CsrfHeader":{"type":"apiKey","in":"header","name":"X-XSRF-TOKEN","description":"The masked token returned by /auth/csrf; matching HttpOnly __Host-LG-CSRF cookie (LG-CSRF in sandbox) is also required."}},
    "parameters":{"AccountId":{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},"Limit":{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":50}},"Offset":{"name":"offset","in":"query","schema":{"type":"integer","minimum":0,"maximum":10000,"default":0}}},
    "responses":{
      "BadRequest":{"description":"Malformed JSON, unknown property, invalid normalized field or pagination","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},
      "Unauthorized":{"description":"Missing/invalid/expired/revoked session, disabled user or invalid credentials","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},
      "Forbidden":{"description":"Role, CSRF or origin boundary rejected the request","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},
      "NotFound":{"description":"Absent or not disclosable to this principal; same public result","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},
      "Throttled":{"description":"Shared authentication budget exhausted","headers":{"Retry-After":{"schema":{"type":"integer","enum":[300]},"description":"Conservative seconds before retry"}},"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}},
      "Unavailable":{"description":"Temporary database/dependency failure; no SQL or credential disclosure","content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/Problem"}}}}
    },
    "schemas":{
      "Minor":{"type":"string","pattern":"^(0|[1-9][0-9]*)$","description":"Exact integer minor units, never JSON floating-point numbers"},
      "Currency":{"type":"string","enum":["CAD","USD","JPY","KWD"]},
      "Registration":{"type":"object","additionalProperties":false,"required":["email","password","displayName"],"properties":{"email":{"type":"string","maxLength":300,"description":"Trimmed/lowercased ASCII email, normalized length at most 254"},"password":{"type":"string","minLength":12,"maxLength":72,"writeOnly":true,"description":"At least 12 characters and at most 72 UTF-8 bytes; no control characters; never silently truncated"},"displayName":{"type":"string","description":"Trimmed value must be 1..80 characters without control characters"}}},
      "Login":{"type":"object","additionalProperties":false,"required":["email","password"],"properties":{"email":{"type":"string","maxLength":300},"password":{"type":"string","minLength":12,"maxLength":72,"writeOnly":true}}},
      "Registered":{"type":"object","additionalProperties":false,"required":["id","email","displayName","role"],"properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string","enum":["CUSTOMER"]}}},
      "Session":{"type":"object","additionalProperties":false,"required":["id","email","displayName","role","expiresAt"],"properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string","enum":["CUSTOMER","ADMIN"]},"expiresAt":{"type":"string","format":"date-time"}}},
      "Csrf":{"type":"object","additionalProperties":false,"required":["headerName","token"],"properties":{"headerName":{"type":"string","enum":["X-XSRF-TOKEN"]},"token":{"type":"string"}}},
      "CreateAccount":{"type":"object","additionalProperties":false,"required":["name","currency"],"properties":{"name":{"type":"string","description":"Trimmed label must contain 1..80 characters without controls"},"currency":{"$ref":"#/components/schemas/Currency"}}},
      "Account":{"type":"object","additionalProperties":false,"required":["id","publicRef","name","currency","postedMinor","reservedMinor","availableMinor","version","updatedAt"],"properties":{"id":{"type":"string","format":"uuid"},"publicRef":{"type":"string"},"name":{"type":"string"},"currency":{"$ref":"#/components/schemas/Currency"},"postedMinor":{"$ref":"#/components/schemas/Minor"},"reservedMinor":{"$ref":"#/components/schemas/Minor"},"availableMinor":{"$ref":"#/components/schemas/Minor"},"version":{"type":"string","pattern":"^[1-9][0-9]*$"},"updatedAt":{"type":"string","format":"date-time"}}},
      "Entry":{"type":"object","additionalProperties":false,"required":["id","journalId","operationId","kind","side","amountMinor","currency","createdAt"],"properties":{"id":{"type":"string"},"journalId":{"type":"string","format":"uuid"},"operationId":{"type":"string","format":"uuid"},"kind":{"type":"string"},"side":{"type":"string","enum":["DEBIT","CREDIT"]},"amountMinor":{"$ref":"#/components/schemas/Minor"},"currency":{"$ref":"#/components/schemas/Currency"},"createdAt":{"type":"string","format":"date-time"}}},
      "Transaction":{"type":"object","additionalProperties":false,"required":["journalId","operationId","kind","effectMinor","currency","createdAt"],"properties":{"journalId":{"type":"string","format":"uuid"},"operationId":{"type":"string","format":"uuid"},"kind":{"type":"string"},"effectMinor":{"type":"string","pattern":"^-?(0|[1-9][0-9]*)$"},"currency":{"$ref":"#/components/schemas/Currency"},"createdAt":{"type":"string","format":"date-time"}}},
      "Recipient":{"type":"object","additionalProperties":false,"required":["publicRef","currency"],"properties":{"publicRef":{"type":"string"},"currency":{"$ref":"#/components/schemas/Currency"}}},
      "SecurityEvent":{"type":"object","additionalProperties":false,"required":["id","actorId","eventType","correlationId","occurredAt"],"properties":{"id":{"type":"string"},"actorId":{"type":"string","format":"uuid","nullable":true},"eventType":{"type":"string"},"correlationId":{"type":"string","format":"uuid"},"occurredAt":{"type":"string","format":"date-time"}}},
      "Problem":{"type":"object","additionalProperties":false,"required":["type","title","status","code","correlationId","validation"],"properties":{"type":{"type":"string"},"title":{"type":"string"},"status":{"type":"integer"},"code":{"type":"string"},"correlationId":{"type":"string","format":"uuid"},"validation":{"type":"object","additionalProperties":{"type":"string"}}}},
      "AccountPage":{"type":"object","required":["items","limit","offset","hasMore"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Account"}},"limit":{"type":"integer"},"offset":{"type":"integer"},"hasMore":{"type":"boolean"}}},
      "EntryPage":{"type":"object","required":["items","limit","offset","hasMore"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Entry"}},"limit":{"type":"integer"},"offset":{"type":"integer"},"hasMore":{"type":"boolean"}}},
      "TransactionPage":{"type":"object","required":["items","limit","offset","hasMore"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}},"limit":{"type":"integer"},"offset":{"type":"integer"},"hasMore":{"type":"boolean"}}},
      "SecurityEventPage":{"type":"object","required":["items","limit","offset","hasMore"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/SecurityEvent"}},"limit":{"type":"integer"},"offset":{"type":"integer"},"hasMore":{"type":"boolean"}}}
    }
  }
}